commonに移した
