--- server/src/main/scala/fpms/calcurator/ldil/LDILMapCalculator.scala
@@ -1,10 +1,0 @@
-package fpms.calcurator.ldil
-
-import fpms.LibraryPackage
-
-trait LDILMapCalculator[F[_]] {
-  def init: F[LDILMap]
-
-  def update(adds: Seq[LibraryPackage]): F[LDILMap]
-
-}
