--- server/src/main/scala/fpms/calcurator/ldil/package.scala
@@ -1,7 +1,0 @@
-package fpms.calcurator
-
-package object ldil {
-  type LDILMap = Map[Int, Seq[Int]]
-
-  private[ldil] val LDIL_REDIS_PREFIX = "directed_"  
-}
