--- server/src/main/scala/fpms/calcurator/rds/RDSContainer.scala
@@ -1,7 +1,0 @@
-package fpms.calcurator.rds
-
-trait RDSContainer[F[_]] {
-  def get(id: Int): F[Option[scala.collection.Set[Int]]]
-
-  def sync(map: RDSMap): F[Unit]
-}
