--- server/src/main/scala/fpms/calcurator/rds/RDSMapCalculator.scala
@@ -1,7 +1,0 @@
-package fpms.calcurator.rds
-
-import fpms.calcurator.ldil.LDILMap
-
-trait RDSMapCalculator[F[_]] {
-  def calc(ldilMap: LDILMap): F[RDSMap]
-}
