--- server/src/main/scala/fpms/calcurator/rds/package.scala
@@ -1,5 +1,0 @@
-package fpms.calcurator
-
-package object rds {
-  type RDSMap = Map[Int, scala.collection.Set[Int]]
-}
