--- server/src/main/scala/fpms/redis/RedisConf.scala
@@ -1,13 +1,0 @@
-package fpms.redis
-
-import com.typesafe.config.Config
-
-case class RedisConf(host: String, port: Int)
-
-object RedisConf {
-  def apply(conf: Config): RedisConf =
-    RedisConf(
-      if (conf.hasPath("host")) conf.getString("host") else "localhost",
-      if (conf.hasPath("port")) conf.getInt("port") else 6379
-    )
-}
