+++ server/src/main/scala/fpms/server/calcurator/ldil/LDILMapCalculator.scala
@@ -1,0 +1,10 @@
+package fpms.server.calcurator.ldil
+
+import fpms.LibraryPackage
+
+trait LDILMapCalculator[F[_]] {
+  def init: F[LDILMap]
+
+  def update(adds: Seq[LibraryPackage]): F[LDILMap]
+
+}
