+++ server/src/main/scala/fpms/server/calcurator/ldil/package.scala
@@ -1,0 +1,7 @@
+package fpms.server.calcurator
+
+package object ldil {
+  type LDILMap = Map[Int, Seq[Int]]
+
+  private[ldil] val LDIL_REDIS_PREFIX = "directed_"  
+}
