+++ server/src/main/scala/fpms/server/calcurator/rds/RDSContainer.scala
@@ -1,0 +1,7 @@
+package fpms.server.calcurator.rds
+
+trait RDSContainer[F[_]] {
+  def get(id: Int): F[Option[scala.collection.Set[Int]]]
+
+  def sync(map: RDSMap): F[Unit]
+}
