+++ server/src/main/scala/fpms/server/calcurator/rds/RDSMapCalculator.scala
@@ -1,0 +1,7 @@
+package fpms.server.calcurator.rds
+
+import fpms.server.calcurator.ldil.LDILMap
+
+trait RDSMapCalculator[F[_]] {
+  def calc(ldilMap: LDILMap): F[RDSMap]
+}
