+++ server/src/main/scala/fpms/server/calcurator/rds/package.scala
@@ -1,0 +1,5 @@
+package fpms.server.calcurator
+
+package object rds {
+  type RDSMap = Map[Int, scala.collection.Set[Int]]
+}
